+++ Runtime/GUIDrawer.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Reflection;
+using Sirenix.OdinInspector;
 #if UNITY_EDITOR
 using UnityEditor;
 #endif
@@ -11,30 +12,50 @@ namespace OcUtility
     public class WorldGUI
     {
         public string text;
-        public int size;
         public Vector3 worldPos;
-        public TextAnchor Alignment = TextAnchor.MiddleCenter;
         public bool drawOnGizmos;
-        public bool dynamicSizing = true;
-        public WorldGUI(string text, Vector3 worldPos, int size)
+        public GUIStyle style;
+        public WorldGUI(string text, Vector3 worldPos, GUIStyle style)
         {
             this.text = text;
-            this.size = size;
+            this.style = style;
             this.worldPos = worldPos;
         }
     }
+    // [ExecuteInEditMode]
     public class GUIDrawer : MonoBehaviour
     {
         public static GUIDrawer Instance => _instance;
         static GUIDrawer _instance;
         public Queue<WorldGUI> guis = new Queue<WorldGUI>();
+        [ShowInInspector][EnumToggleButtons]public HideFlags HideFlags => hideFlags;
         List<WorldGUI> internalGuis = new List<WorldGUI>();
         object _gameView;
+
+
+#if UNITY_EDITOR || DEVELOPMENT_BUILD
+        [InitializeOnLoadMethod]
+        [RuntimeInitializeOnLoadMethod]
+#endif
+        static void Init()
+        {
+            var exists = Resources.FindObjectsOfTypeAll<GUIDrawer>();
+            if(exists.Length > 0)
+            {
+                print("GUI Drawer Exist");
+                return;
+            }
+            
+            var guiDrawer = new GameObject("GUI Drawer (Don'tSave) // OcUtility").AddComponent<GUIDrawer>();
+            // guiDrawer.gameObject.hideFlags = HideFlags.DontSave;
+        }
+        
         void Awake()
         {
 #if UNITY_EDITOR || DEVELOPMENT_BUILD
+            Debug.Log("GUIDrawer Instantiated");
             _instance = this;
-            DontDestroyOnLoad(gameObject);
+            if(Application.isPlaying)DontDestroyOnLoad(gameObject);
 #else
             Destroy(gameObject);
 #endif
@@ -42,6 +63,18 @@ namespace OcUtility
 
         void Update()
         {
+            if(!Application.isPlaying) return;
+            UpdateInternalQueue();
+        }
+
+        void OnDestroy()
+        {
+            if (_instance == this) _instance = null;
+        }
+
+        void UpdateInternalQueue()
+        {
+            if (_instance == null) _instance = this;
             var count = guis.Count;
             if(count == 0) return;
             internalGuis.Clear();
@@ -51,36 +84,32 @@ namespace OcUtility
             }
         }
 
-        void OnDestroy()
+        bool ShouldHideSceneViewGizmo()
         {
-            if (_instance == this) _instance = null;
+#if UNITY_EDITOR
+            if (EditorWindow.focusedWindow == null) return true;
+            
+            if(Application.isPlaying)
+            {
+                var typeName = EditorWindow.focusedWindow.GetType().Name;
+                if (typeName == "GameView") return true;
+            }
+
+            return false;
+#else
+            return true;
+#endif
         }
 
-        bool AreGizmosVisible()
+        static int CalcFontSize(in Vector3 wtsPos, in int size)
         {
-            Assembly asm = Assembly.GetAssembly(typeof(UnityEditor.Editor));
-            Type type = asm.GetType("UnityEditor.GameView");
-            if (_gameView == null)
-            {
-                _gameView = EditorWindow.GetWindow(type);
-            }
-
-            if (_gameView == null) return false;
-            
-            if (type != null)
-            {
-                if (EditorWindow.focusedWindow != SceneView.lastActiveSceneView) return false;
-                // EditorWindow.
-                FieldInfo gizmosField = type.GetField("m_Gizmos", BindingFlags.NonPublic | BindingFlags.Instance);
-                if(gizmosField != null)
-                    return (bool)gizmosField.GetValue(_gameView);
-            }
-            return false;
+            return (int)Mathf.Lerp(size, 5, wtsPos.z / 50);
         }
         
         void OnDrawGizmos()
         {
-            if(AreGizmosVisible()) return;
+            if(ShouldHideSceneViewGizmo()) return;
+            if (!Application.isPlaying) UpdateInternalQueue();
             var count = internalGuis.Count;
             for (int i = 0; i < count; i++)
             {
@@ -92,22 +121,21 @@ namespace OcUtility
                 if(wtsPos.x < 0 || wtsPos.x > Screen.width || 
                    wtsPos.y < 0 || wtsPos.y > Screen.height ||
                    wtsPos.z < 0) continue;
-                
-                var fontSize = gui.dynamicSizing ? (int)Mathf.Lerp(gui.size, 5, wtsPos.z / 50) : gui.size;
+                var rect = new Rect(new Vector2(wtsPos.x, Screen.height - wtsPos.y), Vector2.zero);
+
+                var fontSize = CalcFontSize(in wtsPos, gui.style.fontSize);
+                if (Application.isPlaying) fontSize += 2;
                 if(fontSize < 8) return;
-        
-                var style = new GUIStyle()
-                {
-                    fontSize = fontSize,
-                    richText = true,
-                    alignment = gui.Alignment
-                };
-                Handles.Label(gui.worldPos + new Vector3(0, 0.2f), gui.text, style);
+
+                gui.style.fontSize = fontSize;
+                
+                Handles.Label(gui.worldPos + new Vector3(0, wtsPos.z * 0.075f), gui.text, gui.style);
             }
         }  
         void OnGUI()
         {
             if(Camera.main == null) return;
+            
             var count = internalGuis.Count;
             for (int i = 0; i < count; i++)
             {
@@ -116,18 +144,14 @@ namespace OcUtility
                 if(wtsPos.x < 0 || wtsPos.x > Screen.width || 
                    wtsPos.y < 0 || wtsPos.y > Screen.height ||
                    wtsPos.z < 0) continue;
-                var rect = new Rect(new Vector2(wtsPos.x, Screen.height-wtsPos.y), new Vector2(100, 20));
+                var rect = new Rect(new Vector2(wtsPos.x, Screen.height-wtsPos.y), Vector2.zero);
         
-                var fontSize = gui.dynamicSizing ? (int)Mathf.Lerp(gui.size, 5, wtsPos.z / 50) : gui.size;
+                var fontSize = CalcFontSize(in wtsPos, gui.style.fontSize);
                 if(fontSize < 8) return;
-                
-                var style = new GUIStyle()
-                {
-                    fontSize = fontSize,
-                    richText = true,
-                    alignment = gui.Alignment
-                };
-                GUI.Label(rect, gui.text, style);
+
+                gui.style.fontSize = fontSize;
+             
+                GUI.Label(rect, gui.text, gui.style);
             }
         }
     }